+++ docs/changelog.rst
@@ -8,6 +8,7 @@ CHANGELOG
 **Bug fixes**
 
 - Remove width and height in SVG generating problems in Geotrek-rando V3 by Camille Monchicourt
+- Fix labels filter api v2 (#2764)
 
 **Maintenance**